requirements.txt
alembic==1.18.3
babel==2.9.0
blinker==1.9.0
click==8.3.1
distlib==0.4.0
filelock==3.20.3
flask==3.1.2
flask-migrate==4.1.0
flask-moment==1.0.6
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.3.1
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.3
packaging==26.0
platformdirs==4.5.1
postgres==4.0
psycopg2-binary==2.9.11
psycopg2-pool==1.2
python-dateutil==2.6.0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.46
typing-extensions==4.15.0
virtualenv==20.36.1
werkzeug==3.1.5
wtforms==3.2.1
